Made daytime scene. Minimalist kitchen is best.
Let's face it, this game is gonna be linear. That's okay, because it's not
a traditional video game. It's an interactive narrative.

As a result of this decision, there are no doors in the kitchen until
there needs to be. That means the player is free to explore the kitchen,
but no further. They should focus on making pancakes.